New_Logics

+++ testyanthra/src/Methods/com/Factorial.java
@@ -1,7 +1,7 @@
 package Methods.com;
 
 public class Factorial {
-	public static long factorial(int n) {
+	public static int factorial(int n) {
 		int fact=1;
 		for(int i=1;i<=n;i++) {
 			fact=fact*i;
